arm: keystone: update dt clocks for binding change
Change the bindings for ti,keystone,psc-clock to use a single register.
The previous registers were a combination of the base address and an
offset, which was a bit awkward. Now, reg is just the base address and
the offset will be determined by the power-domain and module-domain
properties.

power-domain is the same as the previous domiain-id property. It needs
a more descriptive name since we are adding another domain property. The
new module-domain property takes the place of one of the reg indexes.

Signed-off-by: David Lechner <david@lechnology.com>