ARM: keystone: make psc clock davinci compatible
The keystone architecture uses a psc virtually identical to that in
the davinci architecture. However, the davinci chips have two psc
controllers, so the static variable base address does not work for
this case.

In order to fix this, I took things one step farther and simplified
the device tree bindings a bit by using reg for the base address and
calculating the offset based on the domain ids instead of using two
addresses that were an odd combinatination of a base address and an
offset.

Device tree documentation and bindings are updated in previous commits.

Signed-off-by: David Lechner <david@lechnology.com>