arm: davinci: da8xx: add 48MHz usb clock
The 48MHz clock for usb11 (ohci) can be derived from usb20 (musb).
This was previously handled in the da8xx ohci usb driver, but that
was the wrong place for such implemenation details. Moving it here
for now.

The plan is to eventually replace it with the common clock framework.