regulator: can change status property for consumers
Add regulator_can_change_status() function for regulator consumers.

This is useful for the da8xx ohci driver which needs to know if it can
turn the regulator on and off.

Also add regulator_is_overcurrent() function.

This is useful for the da8xx ohci driver which needs to know if the
regulator is in an overcurrent state or not.