ARM: davinci: da8xx: update for usb changes
Update da8xx board definitions for changes in the da8xx ohci driver.

The driver now uses a regulator instead of driver-specific platform data.
Currently, there is not a regulator driver that supports an overcurrent
gpio. The plan is to add support for this to the fixed regulator. So,
these will need to be adjusted again when that support is added.