ARM: dts: Add LEGO MINDSTORMS EV3 dts
This adds a device tree definition file for LEGO MINDSTORMS EV3.

What is working:

* Pin muxing
* MicroSD card reader
* UART on input ports 1 and 2
* Buttons
* LEDs
* Poweroff/reset

What is partially working:

* Flash memory - driver loads but can't read the block devices - this is
  probably due to the fact that we are not able to configure the SPI to
  use DMA via device tree
* EEPROM - there seems to be a hardware bug that causes the first byte
  read to be corrupted - this can be worked around by adding an I2C stop
  between writing the register address and reading the data, but the at24
  driver does not have an option to do this

What is not working/to be added later:

* Display - waiting for "tiny DRM" to be mainlined - using spidev for now
* Speaker - needs new PWM sound driver - usable now by gpio/pwm in userspace
* USB - waiting for OHCI and MUSB device tree support to be mainlined
* ADC - needs new iio driver - using spidev for now
* Bluetooth - needs new driver for sequencing power/enable/clock - useable
  now by manipulating gpios/pwm in userspace
* Input and output ports - need some sort of new phy or extcon driver
* Battery - needs new power supply driver (depends on ADC iio driver)

Note on flash partitions:

These partitions are based on the official EV3 firmware from LEGO. It is
expected that most users of the mainline kernel on EV3 will be booting from
an SD card while retaining the official firmware in the flash memory.
Furthermore, the official firmware uses an ancient U-Boot (2009) that has
no device tree support. So, it makes sense to have this parition table in
the EV3 device tree file. In the unlikey case that anyone does create their
own firmware image with different partitioning, they can use a modern
U-Boot in that firmware image that modifies the device tree with the custom
paritions.

Signed-off-by: David Lechner <david@lechnology.com>