ARM: davinci: da850: Fix pwm name matching
This fixes pwm name matching for DA850 familiy devices. When using device
tree, the da850_auxdata_lookup[] table caused pwm devices to have the exact
same name, which caused errors when trying to register the devices.

The names for clock matching in da850_clks[] also have to be updated to
to exactly match in order for the clock lookup to work correctly.

Signed-off-by: David Lechner <david@lechnology.com>