usb: musb: da8xx: fix disconnected not detected in peripheral mode
This fixes disconnection events not being detected when the musb is
set to peripheral mode. We were using a PHY feature to override the
VBUS and ID pins in peripheral mode. However, overriding the VBUS causes
the port to get stuck in a connected state because it always sees a
VBUS good state.

By not overriding the PHY, the VBUS sense is fixed. Also, the ID pin
is internally pulled up in the SoC, so there is no need to override
the ID pin.

Signed-off-by: David Lechner <david@lechnology.com>