Input: pwm-beeper: add optional enable gpio
This adds an optional enable gpio to the pwm-beeper device. This gpio is
used in cases where the beeper needs to be switched on before using. For
example, there may be an amplifier that needs to be enabled.

Tested on LEGO MINDSTORMS EV3, which has a speaker connected to PWM through
an amplifier. The amplifier enable pin is connected to a gpio.

Signed-off-by: David Lechner <david@lechnology.com>