drm/fb: Propogate physical display dimension to fbdev
fbdev has a place for the physical width and height of a display. I would
like to have this information available. This patch works for me, but I
have a strong suspicion that this is not the "right way".

Any suggestions on how to get the proper struct drm_display_info here
rather than assuming the first one is valid.