fbcon: add fbcon=margin:<color> command line option
This adds a new command line option to select the margin color.

The motivation for this is screens where the black does not blend into the
physical surrondings of the screen. For example, using an LCD (not the
backlit kind), white text on a black background is hard to read, so
inverting the colors is preferred. However, the margins are still filled
with black. This makes a big, black, backwards 'L' on the screen. By
setting fbcon=margin:7, the margins white and the LCD looks as expected.

Signed-off-by: David Lechner <david@lechnology.com>