Temporary davinci_clk to enable CONFIG_COMMON_CLK
This converts the clocks in mach-davinci to the common clock framework.

Most of the patch just involves renaming struct clk to struct davinci_clk.
There is also a struct clk_hw added to provide the bridge between the
existing clock implementation and the common clock framework.

The clk_get_parent and clk_set_parent callbacks are dropped because all
clocks currently (effectivly) have a single parent, in which case the
common clock framework does not want you to implement these functions
yourself.

clk_unregister() is dropped because it is not used anywhere in mach-davinci.

EXPORT_SYMBOL() is removed from functions not used outside of mach-davinci.

Signed-off-by: David Lechner <david@lechnology.com>