ARM: da850: remove dummy child clocks
There are a number of child clocks that were created because the clock is
used by more than one device. These clocks previously could not be reused
because it caused linked list issues when we were using a static lookup
table. Now, it is possible to just register extra clkdev device instead of
creating extra clocks.

Signed-off-by: David Lechner <david@lechnology.com>