clk: fix reentrancy of clk_enable() on UP systems
reentrant calls to clk_enable() are not working on UP systems. This is
caused by the fact spin_trylock_irqsave() always returns true when
CONFIG_SMP=n (and CONFIG_DEBUG_SPINLOCK=n) which causes the reference
counting to not work correctly.

This introudces a new set of clk_enable_lock() and clk_enable_unlock()
functions for UP systems that doesn't use spinlocks but effectivly does
the same thing as the SMP version of the functions.

Signed-off-by: David Lechner <david@lechnology.com>