ARM: davinci_all_defconfig: set CONFIG_DAVINCI_WATCHDOG=y
This changes CONFIG_DAVINCI_WATCHDOG from a module to a compiled in
option. Since the reset function has been moved out of the mach code
and into the watchdog driver, devices cannot reset unless the watchdog
driver is loaded, so make it a compiled-in option so that we can always
reboot, even when modules are not loaded.

Suggested-by: Adam Ford <aford173@gmail.com>
Signed-off-by: David Lechner <david@lechnology.com>