ARM: davinci: DA8XX: fix oops in USB PHY driver due to stack allocated platform platform_data
This fixes a possible kernel oops due to using stack allocated platform
data for the USB PHY driver on DA8XX devices. If the platform device
probe is deferred, then we get a corrupt pointer for the platform data.

We now use a global static struct for the platform data so that the
platform data pointer does not get written over.

Fixes: 9b504750923cb ("ARM: davinci: da8xx: Add USB PHY platform device")
Signed-off-by: David Lechner <david@lechnology.com>