ARM: davinci: DA8XX: simplify CFGCHIP regmap_config
Since commit 8253bb3f8255 ("regmap: potentially duplicate the name
string stored in regmap"), the name field of struct regmap_config
is copied in __regmap_init(), so we no longer need to worry about
keeping our own copy of the name.

Just use a string literal in the initalization of da8xx_cfgchip_config
instead of creating a separate variable for the name.

Signed-off-by: David Lechner <david@lechnology.com>