ARM: dts: da850: add power-domains properties to PWM nodes
This adds a power-domains property to the EHRPWM and ECAP nodes in the
TI DA850 device tree. The Linux drivers use this to enable the clocks.
This caused the ECAP PWMs to not work (EHRPWMs worked because the LPSC
for them is a parent of TBCLK and was enabled that way).

Signed-off-by: David Lechner <david@lechnology.com>